* upstream/master: (53 commits)
  AK3+Backend: fix truncated recovery.log with FlashAfterUpdate
  Backend: fixes for booted flashing and addon.d-v2 - only do setup_env and restore_env when in recovery to prevent environment/mount issues post-install - use busybox [ for simple integer test to ensure known numerical operator behavior
  Backend: simplify heredoc usage
  Backend: support for upcoming recovery addon.d-v2 - for FlashAfterUpdate addon.d-v2 support while in recovery - recovery addon.d-v2 requires /system and /system_root stay mounted
  Backend: add additional BOOTCLASSPATH entry for /apex mounts
  Backend: fix /system/bin/* execution on recoveries using LD_CONFIG_FILE
  tools: update magisk utils to v20.4-ed58cf95(20304) canary - fix lzop detection use-case of magiskboot decompress which was broken in the generic stream refactor
  AK3+Backend: mount and output fixes - fix case where TWRP has /system_root mountpoint but ROM is non-SAR - fix regression in ui_print when final character is \
  Backend: work around some Lineage 17 (AOSP Q) Recovery issues - Lineage 17 Recovery has some problems currently: /system mountpoint in ramdisk, neither /system nor /system_root in fstab (it has / instead, which is useless), ANDROID_ROOT=/system, and no /system_root symlink - /etc is linked to /system/etc so once /system is mounted /etc/fstab is gone and no further simple mount commands will work so move other mounts before /system - /system/bin are the only utilities available on PATH with none at all in /sbin so we need /apex files mounted immediately or no non-busybox commands will work after mounting /system over the recovery's skeleton /system - not all devices have /dev/block/bootdevice symlinks, but at some point "broken recovery is broken" has to apply, so I don't intend to build out full system partition discovery for the SAR mount fallback
  AK3+Backend: /postinstall and opposite slot awareness for FlashAfterUpdate addon.d-v2 support - for A/B OTAs my FlashAfterUpdate addon.d script requires the zip to know about addon.d-v2 and act accordingly - signify support with #FLASHAFTERUPDATEV2 tag in updater-script to pass whitelist check since the majority of zips will not work correctly
  Backend: fixes for mounting system on Lineage Recovery - Lineage has a mindboggling: /system_root as mountpoint in ramdisk, neither /system nor /system_root in fstab (it has / instead, which is useless), ANDROID_ROOT=/system, and a possible symlink from /system to /system_root/system - this failure to have a sensible, working fstab triggers the fallback to mounting directly from /dev/block/bootdevice which should exist on all modern devices which would need this SAR fallback - add basic slot support for the fallback mount, since Lineage Recovery also doesn't make symlinks to the active slot partitions
  Backend: put debugging down where it belongs too
  Backend: work around ongoing TWRP SAR support fragmentation TWRP now has 3(!) different competing implementations of SAR currently in use by maintainers: 1) /system as mountpoint in ramdisk/fstab and ANDROID_ROOT=/system, and possible symlink from /system_root to /system 2) /system_root as mountpoint in ramdisk/fstab, ANDROID_ROOT=/system_root, and possible symlink from /system to /system_root/system 3) /system_root as mountpoint in ramdisk/fstab, ANDROID_ROOT=/system, and possible symlink from /system to /system_root/system
  Backend: overdue clean up and refactor - move checks to functions to isolate single-use variables as local - signify globally available variables with capitalization - move setup and mount to functions to mirror umount_all and restore_env - cluster functions by use order (no call should preceed its function) - use short circuit return tests to reduce indentation now that we use functions
  Backend: fix link restoration post-install
  README: document do.systemless, update and sync with XDA thread
  anykernel: add do.systemless
  AK3+Backend: introduce systemless kernel modules support - with do.systemless prop enabled in anykernel.sh, will push the contents of modules/ to create a simple "ak3-helper" Magisk module, where testing on devices thus far shows it loads early enough that even replacement kernel object .ko modules will get loaded - the ak3-helper module info is populated from the kernel.string and parsed kernel version from the kernel being flashed, and if at boot time a different kernel is detected the module automatically removes itself before being loaded to avoid conflicts - since logical partitions and ext4-dedup on Android 10 now prevent mounting /system or /vendor rw under any circumstances, this feature gives kernel developers back the freedom they had before all this to improve kernel .ko modules, .conf files, etc. with the added bonus of it cleaning up after itself
  AK3: add stderr hints for signing debugging
  AK3: clean up and simplify Magisk patch logic - only repatch when there is an Image* (i.e. don't attempt on ramdisk-only mods) - make sure new ramdisk is already copied before trying to detect Magisk patch state in case it changed - clean up Magisk patch state and exported variables immediately
  ...

+++ tools/ak3-core.sh
@@ -15,14 +15,15 @@ split_img=$home/split_img;
 # ui_print "<text>" [...]
 ui_print() {
   until [ ! "$1" ]; do
-    echo -e "ui_print $1\nui_print" > /proc/self/fd/$OUTFD;
+    echo -e "ui_print $1
+      ui_print" >> /proc/self/fd/$OUTFD;
     shift;
   done;
 }
 
 # abort ["<text>" [...]]
 abort() {
-  ui_print " " "$*";
+  ui_print " " "$@";
   exit 1;
 }
 
@@ -160,7 +161,7 @@ unpack_ramdisk() {
     mv -f ramdisk.cpio ramdisk.cpio.$comp;
     $bin/magiskboot decompress ramdisk.cpio.$comp ramdisk.cpio;
     if [ $? != 0 ]; then
-      echo "Attempting with busybox $comp..." >&2;
+      echo "Attempting ramdisk unpack with busybox $comp..." >&2;
       $comp -dc ramdisk.cpio.$comp > ramdisk.cpio;
     fi;
   fi;
@@ -210,10 +211,13 @@ repack_ramdisk() {
   test $? != 0 && packfail=1;
 
   cd $home;
+  $bin/magiskboot cpio ramdisk-new.cpio test;
+  magisk_patched=$?;
+  test $((magisk_patched & 3)) -eq 1 && $bin/magiskboot cpio ramdisk-new.cpio "extract .backup/.magisk $split_img/.magisk";
   if [ "$comp" ]; then
     $bin/magiskboot compress=$comp ramdisk-new.cpio;
     if [ $? != 0 ]; then
-      echo "Attempting with busybox $comp..." >&2;
+      echo "Attempting ramdisk repack with busybox $comp..." >&2;
       $comp -9c ramdisk-new.cpio > ramdisk-new.cpio.$comp;
       test $? != 0 && packfail=1;
       rm -f ramdisk-new.cpio;
@@ -233,7 +237,7 @@ repack_ramdisk() {
 
 # flash_boot (build, sign and write image only)
 flash_boot() {
-  local varlist kernel ramdisk cmdline part0 part1 nocompflag signfail pk8 cert avbtype;
+  local varlist i kernel ramdisk fdt cmdline comp part0 part1 nocompflag signfail pk8 cert avbtype;
 
   cd $split_img;
   if [ -f "$bin/mkimage" ]; then
@@ -304,13 +308,50 @@ flash_boot() {
   else
     test "$kernel" && cp -f $kernel kernel;
     test "$ramdisk" && cp -f $ramdisk ramdisk.cpio;
-    case $kernel in
-      *-dtb) rm -f kernel_dtb;;
-    esac;
     test "$dt" -a -f extra && cp -f $dt extra;
     for i in dtb recovery_dtbo; do
       test "$(eval echo \$$i)" -a -f $i && cp -f $(eval echo \$$i) $i;
     done;
+    case $kernel in
+      *Image*)
+        if [ ! "$magisk_patched" ]; then
+          $bin/magiskboot cpio ramdisk.cpio test;
+          magisk_patched=$?;
+        fi;
+        if [ $((magisk_patched & 3)) -eq 1 ]; then
+          ui_print " " "Magisk detected! Patching kernel so reflashing Magisk is not necessary...";
+          comp=$($bin/magiskboot decompress kernel 2>&1 | grep -v 'raw' | sed -n 's;.*\[\(.*\)\];\1;p');
+          ($bin/magiskboot split $kernel || $bin/magiskboot decompress $kernel kernel) 2>/dev/null;
+          if [ $? != 0 -a "$comp" ]; then
+            echo "Attempting kernel unpack with busybox $comp..." >&2;
+            $comp -dc $kernel > kernel;
+          fi;
+          $bin/magiskboot hexpatch kernel 736B69705F696E697472616D667300 77616E745F696E697472616D667300;
+          if [ "$(file_getprop $home/anykernel.sh do.systemless)" == 1 ]; then
+            strings kernel | grep -E 'Linux version.*#' > $home/vertmp;
+          fi;
+          if [ "$comp" ]; then
+            $bin/magiskboot compress=$comp kernel kernel.$comp;
+            if [ $? != 0 ]; then
+              echo "Attempting kernel repack with busybox $comp..." >&2;
+              $comp -9c kernel > kernel.$comp;
+            fi;
+            mv -f kernel.$comp kernel;
+          fi;
+          test ! -f .magisk && $bin/magiskboot cpio ramdisk.cpio "extract .backup/.magisk .magisk";
+          export $(cat .magisk);
+          test $((magisk_patched & 8)) -ne 0 && export TWOSTAGEINIT=true;
+          for fdt in dtb extra kernel_dtb recovery_dtbo; do
+            test -f $fdt && $bin/magiskboot dtb $fdt patch;
+          done;
+        else
+          case $kernel in
+            *-dtb) rm -f kernel_dtb;;
+          esac;
+        fi;
+        unset magisk_patched KEEPFORCEENCRYPT KEEPVERITY SHA1 TWOSTAGEINIT;
+      ;;
+    esac;
     case $ramdisk_compression in
       none|cpio) nocompflag="-n";;
     esac;
@@ -323,6 +364,7 @@ flash_boot() {
   cd $home;
   if [ -f "$bin/futility" -a -d "$bin/chromeos" ]; then
     if [ -f "$split_img/chromeos" ]; then
+      echo "Signing with CHROMEOS..." >&2;
       $bin/futility vbutil_kernel --pack boot-new-signed.img --keyblock $bin/chromeos/kernel.keyblock --signprivate $bin/chromeos/kernel_data_key.vbprivk --version 1 --vmlinuz boot-new.img --bootloader $bin/chromeos/empty --config $bin/chromeos/empty --arch arm --flags 0x1;
     fi;
     test $? != 0 && signfail=1;
@@ -334,8 +376,9 @@ flash_boot() {
       *recovery*|*SOS*) avbtype=recovery;;
       *) avbtype=boot;;
     esac;
-    if [ "$(/system/bin/dalvikvm -Xbootclasspath:/system/framework/core-oj.jar:/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/bouncycastle.jar -Xnodex2oat -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
-      /system/bin/dalvikvm -Xbootclasspath:/system/framework/core-oj.jar:/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/bouncycastle.jar -Xnodex2oat -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature /$avbtype boot-new.img $pk8 $cert boot-new-signed.img;
+    if [ "$(/system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
+      echo "Signing with AVBv1..." >&2;
+      /system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature /$avbtype boot-new.img $pk8 $cert boot-new-signed.img;
     fi;
   fi;
   if [ $? != 0 -o "$signfail" ]; then
@@ -643,6 +686,9 @@ setup_ak() {
         test "$slot" && slot=_$slot;
       fi;
       if [ "$slot" ]; then
+        if [ -d /postinstall/tmp -a ! "$slot_select" ]; then
+          slot_select=inactive;
+        fi;
         case $slot_select in
           inactive)
             case $slot in
@@ -678,7 +724,9 @@ setup_ak() {
             else
               abort "Unable to determine mtd $block partition. Aborting...";
             fi;
-            target=/dev/mtd/$mtdname;
+            if [ -e /dev/mtd/$mtdname ]; then
+              target=/dev/mtd/$mtdname;
+            fi;
           elif [ -e /dev/block/by-name/$part ]; then
             target=/dev/block/by-name/$part;
           elif [ -e /dev/block/bootdevice/by-name/$part ]; then
@@ -690,7 +738,7 @@ setup_ak() {
           elif [ -e /dev/$part ]; then
             target=/dev/$part;
           fi;
-          test -e "$target" && break 2;
+          test "$target" && break 2;
         done;
       done;
       if [ "$target" ]; then
